Release/0.10.11 rc0

+++ core/services/log/broadcaster.go
@@ -2,10 +2,11 @@ package log
 
 import (
 	"context"
-	"errors"
 	"sync"
 	"sync/atomic"
 	"time"
+
+	"github.com/pkg/errors"
 
 	"github.com/smartcontractkit/chainlink/core/null"
 	"gorm.io/gorm"
@@ -43,20 +44,29 @@ type (
 		utils.DependentAwaiter
 		service.Service
 		httypes.HeadTrackable
+		ReplayFromBlock(number int64)
+
 		IsConnected() bool
 		Register(listener Listener, opts ListenerOpts) (unsubscribe func())
-		LatestHead() *models.Head
-		TrackedAddressesCount() uint32
-		// DB interactions
+
 		WasAlreadyConsumed(db *gorm.DB, lb Broadcast) (bool, error)
 		MarkConsumed(db *gorm.DB, lb Broadcast) error
+		// NOTE: WasAlreadyConsumed and MarkConsumed MUST be used within a single goroutine in order for WasAlreadyConsumed to be accurate
+	}
+
+	BroadcasterInTest interface {
+		Broadcaster
+		BackfillBlockNumber() null.Int64
+		TrackedAddressesCount() uint32
 	}
 
 	broadcaster struct {
-		orm              ORM
-		config           Config
-		connected        *abool.AtomicBool
-		latestHeadFromDb *models.Head
+		orm       ORM
+		config    Config
+		connected *abool.AtomicBool
+
+		// a block number to start backfill from
+		backfillBlockNumber null.Int64
 
 		ethSubscriber *ethSubscriber
 		registrations *registrations
@@ -72,10 +82,14 @@ type (
 		chStop                chan struct{}
 		wgDone                sync.WaitGroup
 		trackedAddressesCount uint32
+		replayChannel         chan int64
+		highestSavedHead      *models.Head
+		lastSeenHeadNumber    int64
 	}
 
 	Config interface {
 		BlockBackfillDepth() uint64
+		BlockBackfillSkip() bool
 		EthFinalityDepth() uint
 		EthLogBackfillBatchSize() uint32
 	}
@@ -109,6 +123,7 @@ var _ Broadcaster = (*broadcaster)(nil)
 // NewBroadcaster creates a new instance of the broadcaster
 func NewBroadcaster(orm ORM, ethClient eth.Client, config Config, highestSavedHead *models.Head) *broadcaster {
 	chStop := make(chan struct{})
+
 	return &broadcaster{
 		orm:              orm,
 		config:           config,
@@ -121,7 +136,8 @@ func NewBroadcaster(orm ORM, ethClient eth.Client, config Config, highestSavedHe
 		newHeads:         utils.NewMailbox(1),
 		DependentAwaiter: utils.NewDependentAwaiter(),
 		chStop:           chStop,
-		latestHeadFromDb: highestSavedHead,
+		highestSavedHead: highestSavedHead,
+		replayChannel:    make(chan int64, 1),
 	}
 }
 
@@ -133,12 +149,20 @@ func (b *broadcaster) Start() error {
 	})
 }
 
-func (b *broadcaster) LatestHead() *models.Head {
-	return b.latestHeadFromDb
+func (b *broadcaster) BackfillBlockNumber() null.Int64 {
+	return b.backfillBlockNumber
 }
 
 func (b *broadcaster) TrackedAddressesCount() uint32 {
 	return atomic.LoadUint32(&b.trackedAddressesCount)
+}
+
+func (b *broadcaster) ReplayFromBlock(number int64) {
+	logger.Infof("LogBroadcaster: Replay requested from block number: %v", number)
+	select {
+	case b.replayChannel <- number:
+	default:
+	}
 }
 
 func (b *broadcaster) Close() error {
@@ -171,17 +195,18 @@ func (b *broadcaster) awaitInitialSubscribers() {
 
 func (b *broadcaster) Register(listener Listener, opts ListenerOpts) (unsubscribe func()) {
 	if len(opts.LogsWithTopics) == 0 {
-		logger.Fatal("Must supply at least 1 Log to Register")
+		logger.Fatal("LogBroadcaster: Must supply at least 1 LogsWithTopics element to Register")
 	}
 
-	wasOverCapacity := b.addSubscriber.Deliver(registration{listener, opts})
+	reg := registration{listener, opts}
+	wasOverCapacity := b.addSubscriber.Deliver(reg)
 	if wasOverCapacity {
-		logger.Error("LogBroadcaster: subscription mailbox is over capacity - dropped the oldest unprocessed subscription")
+		logger.Error("LogBroadcaster: Subscription mailbox is over capacity - dropped the oldest unprocessed subscription")
 	}
 	return func() {
-		wasOverCapacity := b.rmSubscriber.Deliver(registration{listener, opts})
+		wasOverCapacity := b.rmSubscriber.Deliver(reg)
 		if wasOverCapacity {
-			logger.Error("LogBroadcaster: subscription removal mailbox is over capacity - dropped the oldest unprocessed removal")
+			logger.Error("LogBroadcaster: Subscription removal mailbox is over capacity - dropped the oldest unprocessed removal")
 		}
 	}
 }
@@ -222,36 +247,42 @@ func (b *broadcaster) startResubscribeLoop() {
 			return
 		}
 
-		var backfillFrom null.Int64
-		if b.latestHeadFromDb != nil {
+		if b.config.BlockBackfillSkip() && b.highestSavedHead != nil {
+			logger.Warn("LogBroadcaster: BlockBackfillSkip is set to true, preventing a deep backfill - some earlier chain events might be missed.")
+			b.highestSavedHead = nil
+		}
+
+		if b.highestSavedHead != nil {
 			// The backfill needs to start at an earlier block than the one last saved in DB, to account for:
 			// - keeping logs in the in-memory buffers in registration.go
 			//   (which will be lost on node restart) for MAX(NumConfirmations of subscribers)
 			// - HeadTracker saving the heads to DB asynchronously versus LogBroadcaster, where a head
 			//   (or more heads on fast chains) may be saved but not yet processed by LB
 			//   using BlockBackfillDepth makes sure the backfill will be dependent on the per-chain configuration
-			from := b.latestHeadFromDb.Number -
+			from := b.highestSavedHead.Number -
 				int64(b.registrations.highestNumConfirmations) -
 				int64(b.config.BlockBackfillDepth())
-
-			logger.Debugw("LogBroadcaster: Using highest seen head as part of the initial backfill",
-				"blockNumber", b.latestHeadFromDb.Number, "blockHash", b.latestHeadFromDb.Hash,
-				"highestNumConfirmations", b.registrations.highestNumConfirmations, "blockBackfillDepth", b.config.BlockBackfillDepth(),
-			)
-
 			if from < 0 {
 				from = 0
 			}
-			backfillFrom = null.Int64From(from)
+			b.backfillBlockNumber = null.NewInt64(from, true)
+			b.highestSavedHead = nil
 		}
 
-		chBackfilledLogs, abort := b.ethSubscriber.backfillLogs(backfillFrom, addresses, topics)
+		if b.backfillBlockNumber.Valid {
+			logger.Debugw("LogBroadcaster: Using an override as a start of the backfill",
+				"blockNumber", b.backfillBlockNumber.Int64,
+				"highestNumConfirmations", b.registrations.highestNumConfirmations,
+				"blockBackfillDepth", b.config.BlockBackfillDepth(),
+			)
+		}
+
+		chBackfilledLogs, abort := b.ethSubscriber.backfillLogs(b.backfillBlockNumber, addresses, topics)
 		if abort {
 			return
 		}
 
-		// latestHeadFromDb is only used in the first backfill
-		b.latestHeadFromDb = nil
+		b.backfillBlockNumber.Valid = false
 
 		// Each time this loop runs, chRawLogs is reconstituted as:
 		// "remaining logs from last subscription <- backfilled logs <- logs from new subscription"
@@ -268,7 +299,7 @@ func (b *broadcaster) startResubscribeLoop() {
 
 		shouldResubscribe, err := b.eventLoop(chRawLogs, subscription.Err())
 		if err != nil {
-			logger.Warnw("LogBroadcaster: error in the event loop - will reconnect", "err", err)
+			logger.Warnw("LogBroadcaster: Error in the event loop - will reconnect", "err", err)
 			b.connected.UnSet()
 			continue
 		} else if !shouldResubscribe {
@@ -285,10 +316,14 @@ func (b *broadcaster) eventLoop(chRawLogs <-chan types.Log, chErr <-chan error) 
 	debounceResubscribe := time.NewTicker(1 * time.Second)
 	defer debounceResubscribe.Stop()
 
-	logger.Debug("LogBroadcaster: starting the event loop")
+	logger.Debug("LogBroadcaster: Starting the event loop")
 	for {
 		select {
 		case rawLog := <-chRawLogs:
+
+			logger.Debugw("LogBroadcaster: Received a log",
+				"blockNumber", rawLog.BlockNumber, "blockHash", rawLog.BlockHash, "address", rawLog.Address)
+
 			b.onNewLog(rawLog)
 
 		case <-b.newHeads.Notify():
@@ -305,9 +340,14 @@ func (b *broadcaster) eventLoop(chRawLogs <-chan types.Log, chErr <-chan error) 
 		case <-b.rmSubscriber.Notify():
 			needsResubscribe = b.onRmSubscribers() || needsResubscribe
 
+		case blockNumber := <-b.replayChannel:
+			b.backfillBlockNumber.SetValid(blockNumber)
+			logger.Debugw("LogBroadcaster: Returning from the event loop to replay logs from specific block number", "blockNumber", blockNumber)
+			return true, nil
+
 		case <-debounceResubscribe.C:
 			if needsResubscribe {
-				logger.Debug("LogBroadcaster: returning from the event loop to resubscribe")
+				logger.Debug("LogBroadcaster: Returning from the event loop to resubscribe")
 				return true, nil
 			}
 
@@ -318,6 +358,8 @@ func (b *broadcaster) eventLoop(chRawLogs <-chan types.Log, chErr <-chan error) 
 }
 
 func (b *broadcaster) onNewLog(log types.Log) {
+	b.maybeWarnOnLargeBlockNumberDifference(int64(log.BlockNumber))
+
 	if log.Removed {
 		b.logPool.removeLog(log)
 		return
@@ -350,6 +392,8 @@ func (b *broadcaster) onNewHeads() {
 		logger.Debugw("LogBroadcaster: Received head", "blockNumber", latestHead.Number,
 			"blockHash", latestHead.Hash, "parentHash", latestHead.ParentHash, "chainLen", latestHead.ChainLength())
 
+		atomic.StoreInt64(&b.lastSeenHeadNumber, latestHead.Number)
+
 		keptLogsDepth := uint64(b.config.EthFinalityDepth())
 		if b.registrations.highestNumConfirmations > keptLogsDepth {
 			keptLogsDepth = b.registrations.highestNumConfirmations
@@ -361,18 +405,32 @@ func (b *broadcaster) onNewHeads() {
 			keptDepth = 0
 		}
 
-		logs, minBlockNum := b.logPool.getLogsToSend(latestBlockNum)
-
-		if len(logs) > 0 {
-			broadcasts, err := b.orm.FindConsumedLogs(minBlockNum, latestBlockNum)
-			if err != nil {
-				logger.Errorf("Failed to query for log broadcasts, %v", err)
-				return
+		// if all subscribers requested 0 confirmations, we always get and delete all logs from the pool,
+		// without comparing their block numbers to the current head's block number.
+		if b.registrations.highestNumConfirmations == 0 {
+			logs, lowest, highest := b.logPool.getAndDeleteAll()
+			if len(logs) > 0 {
+				broadcasts, err := b.orm.FindConsumedLogs(lowest, highest)
+				if err != nil {
+					logger.Errorf("Failed to query for log broadcasts, %v", err)
+					return
+				}
+				b.registrations.sendLogs(logs, *latestHead, broadcasts)
 			}
+		} else {
+			logs, minBlockNum := b.logPool.getLogsToSend(latestBlockNum)
 
-			b.registrations.sendLogs(logs, *latestHead, broadcasts)
+			if len(logs) > 0 {
+				broadcasts, err := b.orm.FindConsumedLogs(minBlockNum, latestBlockNum)
+				if err != nil {
+					logger.Errorf("LogBroadcaster: Failed to query for log broadcasts, %v", err)
+					return
+				}
+
+				b.registrations.sendLogs(logs, *latestHead, broadcasts)
+			}
+			b.logPool.deleteOlderLogs(uint64(keptDepth))
 		}
-		b.logPool.deleteOlderLogs(uint64(keptDepth))
 	}
 }
 
@@ -387,7 +445,7 @@ func (b *broadcaster) onAddSubscribers() (needsResubscribe bool) {
 			logger.Errorf("expected `registration`, got %T", x)
 			continue
 		}
-		logger.Debugw("LogBroadcaster: Subscribing listener", "requiredBlockConfirmations", reg.opts.NumConfirmations)
+		logger.Debugw("LogBroadcaster: Subscribing listener", "requiredBlockConfirmations", reg.opts.NumConfirmations, "address", reg.opts.Contract)
 		needsResub := b.registrations.addSubscriber(reg)
 		if needsResub {
 			needsResubscribe = true
@@ -407,7 +465,7 @@ func (b *broadcaster) onRmSubscribers() (needsResubscribe bool) {
 			logger.Errorf("expected `registration`, got %T", x)
 			continue
 		}
-		logger.Debugw("LogBroadcaster: Unsubscribing listener", "requiredBlockConfirmations", reg.opts.NumConfirmations)
+		logger.Debugw("LogBroadcaster: Unsubscribing listener", "requiredBlockConfirmations", reg.opts.NumConfirmations, "address", reg.opts.Contract)
 		needsResub := b.registrations.removeSubscriber(reg)
 		if needsResub {
 			needsResubscribe = true
@@ -448,6 +506,20 @@ func (b *broadcaster) appendLogChannel(ch1, ch2 <-chan types.Log) chan types.Log
 	return chCombined
 }
 
+func (b *broadcaster) maybeWarnOnLargeBlockNumberDifference(logBlockNumber int64) {
+	lastSeenHeadNumber := atomic.LoadInt64(&b.lastSeenHeadNumber)
+	diff := logBlockNumber - lastSeenHeadNumber
+	if diff < 0 {
+		diff = -diff
+	}
+
+	if lastSeenHeadNumber > 0 && diff > 1000 {
+		logger.Warnw("LogBroadcaster: Detected a large block number difference between a log and recently seen head. "+
+			"This may indicate a problem with data received from the chain or major network delays.",
+			"lastSeenHeadNumber", lastSeenHeadNumber, "logBlockNumber", logBlockNumber, "diff", diff)
+	}
+}
+
 // WasAlreadyConsumed reports whether the given consumer had already consumed the given log
 func (b *broadcaster) WasAlreadyConsumed(db *gorm.DB, lb Broadcast) (bool, error) {
 	return b.orm.WasBroadcastConsumed(db, lb.RawLog().BlockHash, lb.RawLog().Index, lb.JobID())
@@ -464,8 +536,12 @@ func (n *NullBroadcaster) IsConnected() bool { return false }
 func (n *NullBroadcaster) Register(listener Listener, opts ListenerOpts) (unsubscribe func()) {
 	return func() {}
 }
-func (n *NullBroadcaster) LatestHead() *models.Head {
-	return nil
+
+func (n *NullBroadcaster) ReplayFromBlock(number int64) {
+}
+
+func (n *NullBroadcaster) BackfillBlockNumber() null.Int64 {
+	return null.NewInt64(0, false)
 }
 func (n *NullBroadcaster) TrackedAddressesCount() uint32 {
 	return 0
@@ -476,6 +552,7 @@ func (n *NullBroadcaster) WasAlreadyConsumed(db *gorm.DB, lb Broadcast) (bool, e
 func (n *NullBroadcaster) MarkConsumed(db *gorm.DB, lb Broadcast) error {
 	return errors.New(n.ErrMsg)
 }
+
 func (n *NullBroadcaster) AddDependents(int) {}
 func (n *NullBroadcaster) AwaitDependents() <-chan struct{} {
 	ch := make(chan struct{})
